Rename `TradeCreateScene`->`RampCreateScene`
"Trade" is not specific to fiat on/off ramps and is misleading
